% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_bubble_plot.R
\name{social_bubble_plot}
\alias{social_bubble_plot}
\title{Bubble Plot of Social Media Addiction}
\usage{
social_bubble_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
Bubble Plot of Social Media Addiction
}
